Extract queries into a namespace
