+++ lib/compendium/queries.rb
@@ -1,0 +1,9 @@
+module Compendium
+  module Queries
+    autoload :Query,        'compendium/queries/query'
+    autoload :Collection,   'compendium/queries/collection'
+    autoload :Count,        'compendium/queries/count'
+    autoload :Sum,          'compendium/queries/sum'
+    autoload :Through,      'compendium/queries/through'
+  end
+end
