--- lib/compendium/sum_query.rb
@@ -1,28 +1,0 @@
-require 'compendium/errors'
-require 'compendium/query'
-
-module Compendium
-  # A SumQuery is a Query which runs an SQL sum statement (with a given column)
-  # Often useful in conjunction with a grouped query and counter cache
-  # (alternately, see CountQuery)
-  class SumQuery < Query
-
-    attr_accessor :column
-
-    def initialize(*args)
-      @report = args.shift if arg_is_report?(args.first)
-      @column = args.slice!(1)
-      super(*args)
-
-      @options.reverse_merge!(order: "SUM(#{@column})", reverse: true)
-    end
-
-  private
-
-    def execute_command(command)
-      return [] if command.nil?
-      raise InvalidCommand unless command.respond_to?(:sum)
-      command.sum(column)
-    end
-  end
-end
